Refactor a bit.
