+++ components/PastTenseVerbTable.tsx
